third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-container-white-space-expected.txt

This is an element
 
This is an element
This is an element
  
This is an element
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 1. Was 2.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 2. Was 3.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 3. Was 4.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 4. Was 5.
PASS gFocusedDocument.activeElement.getAttribute("id") is "5"
PASS gFocusedDocument.activeElement.getAttribute("id") is "4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test tests that a anchor element with in image in it and a space after the image, can still give the correct size in ContainerNode::getRect().
